[Rename] 창고 페이지 이름 수정
창고 목록 조회 페이지 와 창고 정보 페이지에 페이지 이름을 변경했습니다.

#70 #71
