+++ frontend/manager/src/views/ContainerDetail.vue
